+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/data/e2ei/MLSConversationsVerificationStatusesHandler.kt
@@ -17,62 +17,21 @@
  */
 package com.wire.kalium.logic.data.e2ei
 
-import com.benasher44.uuid.uuid4
-import com.wire.kalium.cryptography.CryptoCertificateStatus
-import com.wire.kalium.cryptography.WireIdentity
 import com.wire.kalium.logger.KaliumLogger
-import com.wire.kalium.logic.CoreFailure
-import com.wire.kalium.logic.data.client.MLSClientProvider
-import com.wire.kalium.logic.data.conversation.Conversation.VerificationStatus
-import com.wire.kalium.logic.data.conversation.ConversationRepository
-import com.wire.kalium.logic.data.conversation.MLSConversationRepository
-import com.wire.kalium.logic.data.conversation.toModel
-import com.wire.kalium.logic.data.id.ConversationId
-import com.wire.kalium.logic.data.id.GroupID
-import com.wire.kalium.logic.data.id.toDao
-import com.wire.kalium.logic.data.id.toModel
-import com.wire.kalium.logic.data.message.Message
-import com.wire.kalium.logic.data.message.MessageContent
-import com.wire.kalium.logic.data.message.PersistMessageUseCase
-import com.wire.kalium.logic.data.user.UserId
-import com.wire.kalium.logic.data.user.UserRepository
 import com.wire.kalium.logic.data.conversation.EpochChangesObserver
-import com.wire.kalium.logic.functional.Either
-import com.wire.kalium.logic.functional.flatMap
-import com.wire.kalium.logic.functional.getOrElse
-import com.wire.kalium.logic.functional.left
-import com.wire.kalium.logic.functional.map
-import com.wire.kalium.logic.functional.onSuccess
-import com.wire.kalium.logic.functional.right
-import com.wire.kalium.logic.wrapMLSRequest
-import com.wire.kalium.persistence.dao.conversation.EpochChangesDataEntity
-import com.wire.kalium.util.DateTimeUtil
-
-typealias UserToWireIdentity = Map<UserId, List<WireIdentity>>
+import com.wire.kalium.logic.feature.e2ei.usecase.FetchMLSVerificationStatusUseCase
 
 /**
- * Observes all the MLS Conversations Verification status.
- * Notify user (by adding System message in conversation) if needed about changes.
+ * Observes all the MLS Conversations epoch changes and call to re-check Verification status.
  */
 internal interface MLSConversationsVerificationStatusesHandler {
     suspend operator fun invoke()
 }
 
-data class VerificationStatusData(
-    val conversationId: ConversationId,
-    val currentPersistedStatus: VerificationStatus,
-    val newStatus: VerificationStatus
-)
-
 @Suppress("LongParameterList")
 internal class MLSConversationsVerificationStatusesHandlerImpl(
-    private val conversationRepository: ConversationRepository,
-    private val persistMessage: PersistMessageUseCase,
-    private val mlsClientProvider: MLSClientProvider,
-    private val mlsConversationRepository: MLSConversationRepository,
+    private val fetchMLSVerificationStatus: FetchMLSVerificationStatusUseCase,
     private val epochChangesObserver: EpochChangesObserver,
-    private val selfUserId: UserId,
-    private val userRepository: UserRepository,
     kaliumLogger: KaliumLogger
 ) : MLSConversationsVerificationStatusesHandler {
 
@@ -82,149 +41,8 @@ internal class MLSConversationsVerificationStatusesHandlerImpl(
         logger.d("Starting to monitor")
         epochChangesObserver.observe()
             .collect { groupId ->
-
-                mlsClientProvider.getMLSClient()
-                    .flatMap { mlsClient ->
-                        wrapMLSRequest { mlsClient.isGroupVerified(groupId.value) }.map {
-                            it.toModel()
-                        }
-                    }.flatMap { ccGroupStatus ->
-                        if (ccGroupStatus == VerificationStatus.VERIFIED) {
-                            verifyUsersStatus(groupId)
-                        } else {
-                            conversationRepository.getConversationDetailsByMLSGroupId(groupId).map {
-                                VerificationStatusData(
-                                    conversationId = it.conversation.id,
-                                    currentPersistedStatus = it.conversation.mlsVerificationStatus,
-                                    newStatus =
-                                    ccGroupStatus
-                                )
-                            }
-                        }
-                    }.onSuccess {
-                        updateStatusAndNotifyUserIfNeeded(it)
-                    }
+                logger.d("Epoch changed for group $groupId")
+                fetchMLSVerificationStatus(groupId)
             }
-    }
-
-    private suspend fun verifyUsersStatus(groupId: GroupID): Either<CoreFailure, VerificationStatusData> =
-        conversationRepository.getGroupStatusMembersNamesAndHandles(groupId)
-            .flatMap { epochChangesData ->
-                mlsConversationRepository.getMembersIdentities(
-                    epochChangesData.conversationId.toModel(),
-                    epochChangesData.members.keys.map { it.toModel() })
-                    .flatMap { ccIdentities ->
-                        updateKnownUsersIfNeeded(epochChangesData, ccIdentities, groupId)
-                    }
-            }.map { (dbData, ccIdentity) ->
-                var newStatus: VerificationStatus = VerificationStatus.VERIFIED
-                // check that all identities are valid and name and handle are matching
-                for ((userId, wireIdentity) in ccIdentity) {
-                    val persistedMemberInfo = dbData.members[userId.toDao()]
-                    val isUserVerified = wireIdentity.firstOrNull {
-                        it.status != CryptoCertificateStatus.VALID ||
-                                it.certificate == null ||
-                                it.certificate?.displayName != persistedMemberInfo?.name ||
-                                it.certificate?.handle?.handle != persistedMemberInfo?.handle
-                    } == null
-                    if (!isUserVerified) {
-                        newStatus = VerificationStatus.NOT_VERIFIED
-                        break
-                    }
-                }
-                VerificationStatusData(
-                    conversationId = dbData.conversationId.toModel(),
-                    currentPersistedStatus = dbData.mlsVerificationStatus.toModel(),
-                    newStatus = newStatus
-                )
-            }
-
-    private suspend fun updateKnownUsersIfNeeded(
-        epochChangesData: EpochChangesDataEntity,
-        ccIdentities: UserToWireIdentity,
-        groupId: GroupID
-    ): Either<CoreFailure, Pair<EpochChangesDataEntity, UserToWireIdentity>> {
-        var dbData = epochChangesData
-
-        val missingUsers = missingUsers(
-            usersFromDB = epochChangesData.members.keys.map { it.toModel() }.toSet(),
-            usersFromCC = ccIdentities.keys
-        )
-
-        if (missingUsers.isNotEmpty()) {
-            logger.i("Fetching missing users during verification process")
-            conversationRepository.getGroupStatusMembersNamesAndHandles(groupId)
-                .onSuccess {
-                    dbData = it
-                }.getOrElse { error -> return error.left() }
-        }
-
-        return (dbData to ccIdentities).right()
-    }
-
-    private suspend fun missingUsers(usersFromDB: Set<UserId>, usersFromCC: Set<UserId>): Set<UserId> {
-        (usersFromCC - usersFromDB).let {
-            if (it.isNotEmpty()) userRepository.fetchUsersByIds(it)
-            return it
-        }
-    }
-
-    private suspend fun updateStatusAndNotifyUserIfNeeded(
-        verificationStatusData: VerificationStatusData
-    ) {
-        logger.i("Updating verification status and notifying user if needed")
-        val newStatus = getActualNewStatus(
-            newStatusFromCC = verificationStatusData.newStatus,
-            currentStatus = verificationStatusData.currentPersistedStatus
-        )
-
-        if (newStatus == verificationStatusData.currentPersistedStatus) return
-
-        conversationRepository.updateMlsVerificationStatus(newStatus, verificationStatusData.conversationId)
-
-        if (newStatus == VerificationStatus.DEGRADED || newStatus == VerificationStatus.VERIFIED) {
-            notifyUserAboutStateChanges(verificationStatusData.conversationId, newStatus)
-        }
-    }
-
-    /**
-     * Current CoreCrypto implementation returns only a boolean flag "if conversation is verified or not".
-     * So we need to calculate if status was degraded on our side by comparing it to the previous status.
-     * @param newStatusFromCC - [VerificationStatus] that CoreCrypto returns.
-     * @param currentStatus - [VerificationStatus] that conversation currently has.
-     * @return [VerificationStatus] that should be saved to DB.
-     */
-    private fun getActualNewStatus(newStatusFromCC: VerificationStatus, currentStatus: VerificationStatus): VerificationStatus =
-        if (newStatusFromCC == VerificationStatus.NOT_VERIFIED && currentStatus == VerificationStatus.VERIFIED)
-            VerificationStatus.DEGRADED
-        else if (newStatusFromCC == VerificationStatus.NOT_VERIFIED && currentStatus == VerificationStatus.DEGRADED)
-            VerificationStatus.DEGRADED
-        else
-            newStatusFromCC
-
-    /**
-     * Add a SystemMessage into a conversation, to inform user when the conversation verification status becomes DEGRADED.
-     */
-    private suspend fun notifyUserAboutStateChanges(
-        conversationId: ConversationId,
-        updatedStatus: VerificationStatus
-    ) {
-        logger.i("Notifying user about state changes")
-        val content = if (updatedStatus == VerificationStatus.VERIFIED) MessageContent.ConversationVerifiedMLS
-        else MessageContent.ConversationDegradedMLS
-
-        val conversationDegradedMessage = Message.System(
-            id = uuid4().toString(),
-            content = content,
-            conversationId = conversationId,
-            date = DateTimeUtil.currentIsoDateTimeString(),
-            senderUserId = selfUserId,
-            status = Message.Status.Sent,
-            visibility = Message.Visibility.VISIBLE,
-            expirationData = null
-        )
-
-        persistMessage(conversationDegradedMessage)
-        conversationRepository.setDegradedConversationNotifiedFlag(conversationId, updatedStatus != VerificationStatus.DEGRADED)
     }
 }
