+++ logic/src/commonTest/kotlin/com/wire/kalium/logic/util/arrangement/usecase/FetchMLSVerificationStatusArrangement.kt
@@ -1,0 +1,43 @@
+/*
+ * Wire
+ * Copyright (C) 2024 Wire Swiss GmbH
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see http://www.gnu.org/licenses/.
+ */
+package com.wire.kalium.logic.util.arrangement.usecase
+
+import com.wire.kalium.logic.feature.e2ei.usecase.FetchMLSVerificationStatusUseCase
+import io.mockative.Mock
+import io.mockative.any
+import io.mockative.given
+import io.mockative.mock
+
+internal interface FetchMLSVerificationStatusArrangement {
+
+    val fetchMLSVerificationStatusUseCase: FetchMLSVerificationStatusUseCase
+}
+
+internal class FetchMLSVerificationStatusArrangementImpl : FetchMLSVerificationStatusArrangement {
+
+    @Mock
+    override val fetchMLSVerificationStatusUseCase: FetchMLSVerificationStatusUseCase =
+        mock(FetchMLSVerificationStatusUseCase::class)
+
+    init {
+        given(fetchMLSVerificationStatusUseCase)
+            .suspendFunction(fetchMLSVerificationStatusUseCase::invoke)
+            .whenInvokedWith(any())
+            .thenReturn(Unit)
+    }
+}
